More fixes for building on GCC 4.3, made archopt work with newer CPUs with many flags.
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@5234 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/tools/depbuild.cpp
@@ -236,7 +236,7 @@ int main(int argc, const char *const *const argv)
     else
     {
       struct stat unused;
-      for (size_t i = 5; i < argc; i++)
+      for (int i = 5; i < argc; i++)
       {
         dependency_calculate(argv[i], unused);
       }
